requirements.txt
fastapi==0.135.1
uvicorn==0.34.2
asyncpg==0.30.0
requests==2.32.3
python-dotenv==1.0.1
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
bcrypt<5
pytest==9.0.3
httpx==0.28.1
